Lesson-3-After-The-Basics/VMs-In-Azure/requirements.txt
azure-mgmt-cpute
azure-cli-core